publisher/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
click==8.1.3
    # via -r requirements.in
numpy==1.24.1
    # via -r requirements.in
paho-mqtt==1.6.1
    # via -r requirements.in
pydantic==1.10.4
    # via -r requirements.in
typing-extensions==4.4.0
    # via pydantic
